extract search-related reducer / actions
